Refactor server configuration
