+++ server/routerlicious/packages/lambdas/src/deli/lambdaFactory.ts
@@ -15,11 +15,12 @@ import {
     IPartitionLambda,
     IPartitionLambdaFactory,
     IProducer,
+    IServiceConfiguration,
     ITenantManager,
     MongoManager,
 } from "@fluidframework/server-services-core";
 import { generateServiceProtocolEntries } from "@fluidframework/protocol-base";
-import { FileMode, IServiceConfiguration } from "@fluidframework/protocol-definitions";
+import { FileMode } from "@fluidframework/protocol-definitions";
 import { IGitManager } from "@fluidframework/server-services-client";
 import { Provider } from "nconf";
 import { NoOpLambda } from "../utils";
