+++ server/routerlicious/packages/memory-orderer/src/remoteNode.ts
@@ -6,8 +6,15 @@
 import assert from "assert";
 import { EventEmitter } from "events";
 import { Deferred } from "@fluidframework/common-utils";
-import { IClient, IDocumentMessage, IServiceConfiguration } from "@fluidframework/protocol-definitions";
-import { INode, IOrderer, IOrdererConnection, IWebSocket, MongoManager } from "@fluidframework/server-services-core";
+import { IClient, IDocumentMessage } from "@fluidframework/protocol-definitions";
+import {
+    INode,
+    IOrderer,
+    IOrdererConnection,
+    IServiceConfiguration,
+    IWebSocket,
+    MongoManager,
+} from "@fluidframework/server-services-core";
 import { debug } from "./debug";
 import { IConcreteNode, IConnectedMessage, IConnectMessage, INodeMessage, IOpMessage } from "./interfaces";
 import { IOrdererConnectionFactory, ProxyOrderer } from "./proxyOrderer";
