+++ server/routerlicious/packages/services-core/src/index.ts
@@ -9,6 +9,7 @@ export * from "./clientManager";
 export * from "./combinedContext";
 export * from "./combinedLambda";
 export * from "./combinedProducer";
+export * from "./configuration";
 export * from "./database";
 export * from "./document";
 export * from "./emptyTaskMessageSender";
